refactor(dashboard): extract the conversation rebind stack section
Fifth composition-section extraction (B1): the conversation comment/bookmark
stores, the session rebind coordinator, the viewer store adapters, and the
rebind restore task move into src/dashboard/sections/conversationStack.ts.
The section call sits at the coordinator's original position so the
shared-authority wiring keeps its exact textual shape (the architecture
guard pins it) and the bootstrap tick sequence is unchanged. No behavior
change.
